Update README.md
An Overview of the Analysis:

We dived into the world of cryptocurrencies in this detailed investigation, utilising data visualisation techniques to extract useful insights from a massive dataset spanning 2013 to 2021. We created a dynamic dashboard using Microsoft Power BI that displayed the historical values and market capitalizations of popular cryptocurrencies such as BTC, ETH, and XRP, among others. The visualisations, which included area charts, bar graphs, pie charts, and line charts, shed light on the bitcoin market's growth trajectory, emphasising variations and future patterns. The correlation coefficient card, which revealed the interaction between different cryptocurrencies, was a unique feature. The slicer tool now has an interactive dimension that allows users to search for certain coins. 

This analysis, based on Tufte's effectiveness metric, provides a comprehensive view of the cryptocurrency ecosystem, assisting investors in comprehending market dynamics and prospective investment possibilities while emphasising the volatile character of this digital asset class.